+++ src/eca/features/tools.clj
@@ -116,12 +116,15 @@
                                                                         (assoc :type :mcp)
                                                                         (update :tools #(mapv with-tool-status %)))))})))
 
-(defn manual-approval? [name config]
+(defn manual-approval? [all-tools name args db config]
   (boolean
-   (let [manual-approval? (get-in config [:toolCall :manualApproval] nil)]
-     (if (coll? manual-approval?)
-       (some #(= name (str %)) manual-approval?)
-       manual-approval?))))
+   (let [require-approval-fn (:require-approval-fn (first (filter #(= name (:name %))
+                                                                  all-tools)))
+         manual-approval? (get-in config [:toolCall :manualApproval] nil)]
+     (or (when require-approval-fn (require-approval-fn args {:db db}))
+         (if (coll? manual-approval?)
+           (some #(= name (str %)) manual-approval?)
+           manual-approval?)))))
 
 (defn tool-call-summary [all-tools name args]
   (when-let [summary-fn (:summary-fn (first (filter #(= name (:name %))